fix: workflow always sending discord notifs
The conditional statements were all in the form of `input || true`,
presumably in an effort to treat this as `true` if the input was
not set as it would be when the trigger is a push rather than a manual
dispatch (`null || true` => `true`). However `false || true` also
results in `true`, so this would just always result in `true`.

Changing this to `!= false` does result in the desired behaviour:

- Enabled: `true != false` => `true`
- Disabled: `false != false` => `false`
- Unset: `null != false` => `true`